Update and rename timer.c to timer_uWave.c
